refactor : BattleResult 디렉토리 이동 #PAR-340
